requirements.txt
fastapi
uvicorn
PyPDF2
openai
azure-search-documents
azure-core
python-dotenv
